services/drift/requirements.txt
fastapi
uvicorn[standard]
prometheus_client
psycopg2-binary
sqlalchemy
numpy
pandas
python-dotenv
